--- src/datasets/utils.py
@@ -1,34 +1,0 @@
-import importlib
-from datasets.base_dataset import BaseDataset
-
-
-def contains_key(opt, key):
-    try:
-        _ = opt[key]
-        return True
-    except:
-        return False
-
-
-def find_dataset_using_name(dataset_name):
-    """Import the module "data/[dataset_name]_dataset.py".
-    In the file, the class called DatasetNameDataset() will
-    be instantiated. It has to be a subclass of BaseDataset,
-    and it is case-insensitive.
-    """
-    dataset_filename = "datasets." + dataset_name + "_dataset"
-    datasetlib = importlib.import_module(dataset_filename)
-
-    dataset = None
-    target_dataset_name = dataset_name.replace("_", "") + "dataset"
-    for name, cls in datasetlib.__dict__.items():
-        if name.lower() == target_dataset_name.lower() and issubclass(cls, BaseDataset):
-            dataset = cls
-
-    if dataset is None:
-        raise NotImplementedError(
-            "In %s.py, there should be a subclass of BaseDataset with class name that matches %s in lowercase."
-            % (dataset_filename, target_dataset_name)
-        )
-
-    return dataset
